mjd_parameter.py
